+++ tibiawikisql/__main__.py
@@ -35,6 +35,18 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 import collections
 import datetime
 import json
@@ -116,7 +128,7 @@ categories = {
     "house": Category("Player-Ownable Buildings", parsers.HouseParser, no_images=True),
     "charm": Category("Charms", parsers.CharmParser, extension=".png"),
     "outfits": Category("Outfits", parsers.OutfitParser, no_images=True),
-    # "worlds": Category("Game Worlds", models.World, no_images=True, include_deprecated=True),
+    "worlds": Category("Game Worlds", parsers.WorldParser, no_images=True, include_deprecated=True),
     # "mounts": Category("Mounts", models.Mount),
     # "updates": Category("Updates", models.Update, no_images=True),
 }