+++ tibiawikisql/models/world.py
@@ -1,5 +1,4 @@
-#
-#  Copyright 2021 Allan Galarza
+#  Copyright (c) 2025.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -13,49 +12,55 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from tibiawikisql import schema
-from tibiawikisql.models import abc
-from tibiawikisql.utils import parse_boolean, parse_integer, parse_date
+#
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+import datetime
+
+from tibiawikisql.api import WikiEntry
+from tibiawikisql.utils import parse_boolean, parse_date, parse_integer
 
 
-class World(abc.Row, abc.Parseable, table=schema.World):
-    """Represents a Game World.
+class World(WikiEntry):
+    """Represents a Game World."""
 
-    Attributes
-    ----------
-    article_id: :class:`int`
-        The id of the containing article.
-    title: :class:`str`
-        The title of the containing article.
-    timestamp: :class:`int`
-        The last time the containing article was edited.
-    name: :class:`str`
-        The name of the world.
-    pvp_type: :class:`str`
-        The world's PvP type.
-    location: :class:`str`
-        The world's server's physical location.
-    preview: :class:`bool`
-        Whether the world is a preview world or not.
-    experimental: :class:`bool`
-        Whether the world is a experimental world or not.
-    online_since: :class:`str`
-        Date when the world became online for the first time, in ISO 8601 format.
-    offline_since: :class:`str`, optional
-        Date when the world went offline, in ISO 8601 format.
-    merged_into: :class:`str`, optional.
-        The name of the world this world got merged into, if applicable.
-    battleye: :class:`bool`
-        Whether the world is BattlEye protected or not.
-    battleye_type: :class:`bool`
-        The type of BattlEye protection the world has. Can be either green or yellow.
-    protected_since: :class:`str`
-        Date when the world started being protected by BattlEye, in ISO 8601 format.
-    world_board: :class:`int`
-        The board ID for the world's board.
-    trade_board: :class:`int`
-        The board ID for the world's trade board.
-    """
+    name: str
+    """The name of the world."""
+    pvp_type: str
+    """The world's PvP type."""
+    location: str
+    """The world's server's physical location."""
+    is_preview: bool
+    """Whether the world is a preview world or not."""
+    is_experimental: bool
+    """Whether the world is a experimental world or not."""
+    online_since: datetime.date
+    """Date when the world became online for the first time, in ISO 8601 format."""
+    offline_since: datetime.date | None
+    """Date when the world went offline, in ISO 8601 format."""
+    merged_into: str | None
+    """The name of the world this world got merged into, if applicable."""
+    battleye: bool
+    """Whether the world is BattlEye protected or not."""
+    battleye_type: str | None
+    """The type of BattlEye protection the world has. Can be either green or yellow."""
+    protected_since: datetime.date | None
+    """Date when the world started being protected by BattlEye, in ISO 8601 format."""
+    world_board: int | None
+    """The board ID for the world's board."""
+    trade_board: int | None
+    """The board ID for the world's trade board."""
+
 
     _map = {
         "name": ("name", str.strip),